requirements.txt
fire
loguru
manga-ocr
easyocr
natsort
numpy
opencv-python
Pillow
pyclipper
pytest
requests
scipy
shapely
torch
torchsummary
tqdm
yattag
transformers
datasets
PyQt5
